int en long int

+++ include/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: bifrah <bifrah@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/11 16:24:45 by bifrah            #+#    #+#             */
-/*   Updated: 2021/12/17 16:22:02 by bifrah           ###   ########.fr       */
+/*   Updated: 2021/12/17 16:25:48 by bifrah           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,17 +27,17 @@ typedef struct s_param {
 	int			j;
 }				t_param;
 
-void	print_tab(long int *tab, int size);
-int		ft_issign(int c);
-int		ptrstrlen(char **dest);
-void	ft_freetab(long int **tab);
-void	ft_free(char **tmp, char ***dest);
-int		ft_stockandcheck(int argc, char **argv);
-int		ft_argisvalid(char *argv);
-int		*argv_to_tab(char **argv, int size);
-int		*check_format(long int *tab, int size);
-int		ft_checkdouble(long int *tab, int size);
-int		ft_maxandmin(long int *tab, int size);
+void		print_tab(long int *tab, int size);
+int			ft_issign(int c);
+int			ptrstrlen(char **dest);
+void		ft_freetab(long int **tab);
+void		ft_free(char **tmp, char ***dest);
+int			ft_stockandcheck(int argc, char **argv);
+int			ft_argisvalid(char *argv);
+long int	*argv_to_tab(char **argv, int size);
+long int	*check_format(long int *tab, int size);
+int			ft_checkdouble(long int *tab, int size);
+int			ft_maxandmin(long int *tab, int size);
 
 //#define $BR exit (0);
 #define ERROR			-2
